refactor: Move typescript types to declarations files
- Move redux types into store.d.ts
- Move others to more specific declarations or global

Closes PFMS-56.
