--- client/global.d.ts
@@ -1,4 +1,0 @@
-declare module '*.scss' {
-  const content: {[className: string]: string};
-  export default content;
-}
