+++ client/reducers/index.ts
@@ -1,12 +1,7 @@
 import { combineReducers, Reducer } from 'redux';
 import { combineEpics, Epic } from 'redux-observable';
-import user, { userEpic, UserState } from './user';
-import dropdown, { DropdownState } from './dropdown';
-
-export interface AppState {
-  user: UserState;
-  dropdown: DropdownState;
-}
+import user, { userEpic } from './user';
+import dropdown from './dropdown';
 
 export const rootEpic: Epic = combineEpics<Epic>(
   userEpic,
