+++ client/shared/global.d.ts
@@ -1,0 +1,20 @@
+declare module '*.scss' {
+  const content: {[className: string]: string};
+  export default content;
+}
+
+interface Window {
+  __REDUX_DEVTOOLS_EXTENSION__?: any | Function;
+}
+
+type DeepReadonly<T> =
+  T extends (infer R)[] ? DeepReadonlyArray<R> :
+    T extends Function ? T :
+      T extends object ? DeepReadonlyObject<T> :
+        T;
+
+interface DeepReadonlyArray<T> extends ReadonlyArray<DeepReadonly<T>> {}
+
+type DeepReadonlyObject<T> = {
+  readonly [P in keyof T]: DeepReadonly<T[P]>;
+};
