+++ client/shared/portal.d.ts
@@ -1,0 +1,24 @@
+interface PortalRect {
+  top?: number | string;
+  right?: number | string;
+  bottom?: number | string;
+  left?: number | string;
+  width?: number | string;
+  height?: number | string;
+  [key: string]: number | string | undefined;
+}
+
+interface Style {
+  [propName: string]: string;
+}
+
+interface Item {
+  text: string;
+  value: string;
+  style?: Style;
+}
+
+interface ItemSelection {
+  data?: string;
+  time?: string;
+}
