+++ client/shared/portfolio.d.ts
@@ -1,0 +1,9 @@
+interface Portfolio {
+  name: string;
+  brokerage: string;
+  balance: number;
+  returns: number;
+  isGroup: boolean;
+  isRetirement: boolean;
+  isSavings: boolean;
+}
