--- client/shared/styleProps.ts
@@ -1,11 +1,0 @@
-export interface ClientRect {
-  top?: number | string;
-  right?: number | string;
-  bottom?: number | string;
-  left?: number | string;
-  width?: number | string;
-  height?: number | string;
-  [key: string]: string | number | undefined;
-}
-
-export default {};
